Create Knapsack
This code implements the Fractional Knapsack Problem using a Greedy Algorithm. The objective is to maximize the total value of items that can be placed in a knapsack with a given capacity
This approach provides a simple solution to the Fractional Knapsack Problem, where items can be taken in fractions, and the goal is to maximize the value of the knapsack.